+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* `server_agent` field in activity webhook payload using `netius.IDENTIFIER`
-* `contents_size` field in activity webhook payload
+*
 
 ### Changed
 
@@ -19,6 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+
+## [1.43.0] - 2026-04-12
+
+### Added
+
+* `server_agent` field in activity webhook payload using `netius.IDENTIFIER`
+* `contents_size` field in activity webhook payload
 
 ## [1.42.0] - 2026-04-12